Created Navbar directory and styled component. Used in ParticipantShowPage and StaffDashboard components.
