+++ app/javascript/components/Navbar/styles.js
@@ -1,0 +1,29 @@
+/*
+ * Navbar Styles
+ *
+ * This contains all the styles for the Navbar container.
+ */
+
+export const styles = theme => ({
+  navBar: {
+    maxHeight: '100%',
+    paddingTop: 40,
+    backgroundColor: theme.palette.common.black,
+  },
+  navBarSignOut: {
+    color: theme.palette.common.white,
+  },
+  navBarItem: {
+    color: theme.palette.common.white,
+  },
+  unloopLogo: {
+    paddingLeft: '10px',
+    paddingBottom: '10px',
+    width: '100%',
+    objectFit: 'contain',
+    overflowX: 'hidden',
+    backgroundColor: theme.palette.common.black,
+  },
+});
+
+export default styles;
